Update values in balance popover
